Re #618 Even more robust migration of repeat type

+++ grails-app/services/us/wearecurio/services/MigrationService.groovy
@@ -515,9 +515,11 @@ class MigrationService {
 			sql('alter table user_group drop index full_name')
 		}
 		tryMigration("Change repeat type column") {
-			if (!sql('alter table entry change repeat_type repeat_type_id int(11)')) {
-				sql('alter table entry drop column repeat_type_id')
-				sql('alter table entry change repeat_type repeat_type_id int(11)')
+			if (sql('select * from entry where repeat_type is null limit 1')) {
+				if (!sql('alter table entry change repeat_type repeat_type_id int(11)')) {
+					sql('alter table entry drop column repeat_type_id')
+					sql('alter table entry change repeat_type repeat_type_id int(11)')
+				}
 			}
 		}
 	}
